Restructure project  - rename src folder to app  - create app.js
